Fix three compiler errors from net8.0 build
- ProductionPatternElement: cast TokenId enum to int before multiply
  in GetHashCode (enums require explicit cast for arithmetic)
- MemberElement: Next field and Link parameter are MemberElement?
  (tail element legitimately links to null)
- InvocationListElement: nextElement is MemberElement?; suppress
  null-forgiving on IList object? casts (elements always MemberElement)

+++ src/Yale/Parser/ProductionPatternElement.cs
@@ -108,7 +108,7 @@ internal sealed class ProductionPatternElement
         }
     }
 
-    public override int GetHashCode() => Id * 37;
+    public override int GetHashCode() => (int)Id * 37;
 
     public override string ToString()
     {
